small fixed and peekaboo
not gonna be on the jam page until tomorrow. If you guys want to fix anything else please do. If you have a problem with peekaboo existing we talk it over ig. But I don't see a reason not to just give it to the player tbh. also den if you wanna make the .bps for this pls add "V1.1" to the mod string thx